transport_plugins/bled112/blob/requirements.txt
iotile-core>=5.0.0,<6
pyserial>=3.4.0,<4
iotile_transport_bled112